pyproject.toml
[tool.poetry]
name = "xi.auth"
version = "0.2.0"
description = "Backend for user-related services"
authors = ["niqzart", "xi.team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.18"}
pydantic-marshals = "0.3.12"
aenum = "^3.1.15"
passlib = "^1.7.4"
alembic = "^1.12.1"
aio-pika = "^9.3.1"
python-multipart = "^0.0.9"
filetype = "^1.2.0"
cryptography = "^42.0.5"
discord-webhook = {extras = ["async"], version = "^1.3.1"}
aiogram = "^3.4.1"
httpx = "^0.23.0"
psycopg = {extras = ["binary"], version = "^3.1.19"}
redis = {extras = ["hiredis"], version = "^5.1.1"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
flake8 = "4.0.1"
black = "^24.3.0"
pre-commit = "^3.2.1"
mypy = "^1.1.1"
flake8-pie = "0.16.0"
dlint = "0.14.0"
flake8-coding = "1.3.2"
flake8-secure-coding-standard = "1.4.0"
flake8-unused-arguments = "0.0.13"
flake8-pytest-style = "1.7.2"
flake8-useless-assert = "0.4.4"
flake8-simplify = "0.19.3"
flake8-future-import = "0.4.7"
flake8-functions-names = "0.4.0"
flake8-variables-names = "0.0.5"
flake8-comments = "0.1.2"
flake8-annotations = "2.9.1"
flake8-new-union-types = "0.4.1"
flake8-pep585 = "0.1.7"
flake8-type-checking = "2.3.1"
flake8-print = "5.0.0"
flake8-use-fstring = "1.4"
flake8-use-pathlib = "0.3.0"
flake8-test-name = "^0.1.5"
wemake-python-styleguide = "^0.17.0"
types-passlib = "^1.7.7.13"
pydantic-marshals = {extras = ["assert-contains"], version = "^0.3.5"}
faker = "^19.6.2"
freezegun = "^1.2.2"
coverage = "^7.3.2"
pytest-cov = "^4.1.0"
pillow = "9.5.0"
faker-file = "^0.17.11"
rstr = "3.2.2"
watchfiles = "^0.21.0"
polyfactory = "^2.15.0"

[tool.isort]
profile = "black"
py_version = 311
combine_as_imports = true
known_first_party = ["app", "supbot", "tests"]
no_lines_before = "LOCALFOLDER"
reverse_relative = true
line_length = 88

[tool.mypy]
strict = true
exclude = ["alembic"]
plugins = "pydantic_marshals.mypy.plugin"

[tool.coverage.run]
concurrency = ["greenlet", "thread"]
sigterm = true

[tool.coverage.report]
show_missing = true
skip_covered = true

fail_under = 100

exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
  "if sys.platform .*:",
  "if( not)? PRODUCTION_MODE",
  "except ImportError",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == [\"']__main__[\"']:",
  "class .*\\(Protocol\\):",
  "@abstractmethod",
  "custom_swagger_ui_html",
  "@mark.skip",
  "!wow!",
]

omit = [
  "alembic/*",
  "test/*",
  "app/common/fastapi_ext.py",
  "app/common/sqlalchemy_ext.py",
  "app/common/starlette_cors_ext.py",
  "app/supbot/setup.py",
  "app/supbot/utils/aiogram_ext.py",
  "app/setup_ci.py",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
